Door Script updated for more keys, (some of) Katherine Final Models added
Can now pick what type of key goes where, can be selected based on shape in inspector (going off the current greyboxes that have a triangle, circle, and square key). If the player approaches a door it says which key is needed, if the player has the right key it prompts to use that key, otherwise stays the same. Each key needs to be tagged with what shape it is to work, this will update the game manager that it has been collected, which will allow the door to work. Circle prefab already works.
Ladybug exhibit, spiketrap, gem update, final wall lamps added.